feat: nethsecurity, add hardware_table
Creation of a model representign hardware data from the database,
creation of a controller to manage the logic of operations to be
performed on the hardware, creation of a web route associated with
the controller, and creation of a view for data representation to users.
Finally, a slight modification in the welcome view for adding a link that
leads to the hardware web route.